+++ README.md
@@ -1,5 +1,7 @@
 ## To do list by React
 
+### node v16.17.0
+
 実行するために下記のコマンドを入力してください。
 
 - npm install
