+++ src/components/molecules/CountsTodo.jsx
@@ -5,17 +5,19 @@ export const CountsTodo = (props) => {
   const contentStyle = {
     margin: "10px",
   };
+  const eachTodoCount = [
+    { label: "全てのタスク", color: "gold", count: wholeCounts },
+    { label: "完了済", color: "aqua", count: doneCounts },
+    { label: "未完了", color: "salmon", count: doneYetCounts },
+  ];
+
   return (
     <div style={contentStyle}>
-      <CountTodo countColor={"gold"} countNum={wholeCounts}>
-        全てのタスク
-      </CountTodo>
-      <CountTodo countColor={"aqua"} countNum={doneCounts}>
-        完了済
-      </CountTodo>
-      <CountTodo countColor={"salmon"} countNum={doneYetCounts}>
-        未完了
-      </CountTodo>
+      {eachTodoCount.map((todoCount) => (
+        <CountTodo countColor={todoCount.color} countNum={todoCount.count}>
+          {todoCount.label}
+        </CountTodo>
+      ))}
     </div>
   );
 };
